backend/requirements.txt
fastapi
uvicorn[standard]
jinja2
aiofiles
geopandas
rasterio
shapely
numpy
pillow
opencv-python-headless
segmentation-models-pytorch
sqlalchemy
python-multipart